Rearranged whole prophy package. Refactored and moved the code (yes, you know what's still passin').
