--- prophy/base.py
@@ -1,14 +1,0 @@
-
-class kind(type):
-    INT = ('INT', 0)
-    ENUM = ('ENUM', 1)
-    BYTES = ('BYTES', 2)
-    ARRAY = ('ARRAY', 3)
-    STRUCT = ('STRUCT', 4)
-    UNION = ('UNION', 5)
-
-
-class prophy_data_object(object):
-    _is_prophy_object = True
-    __slots__ = []
-
